+++ examples/data-sources/elasticstack_kibana_security_role/data-source.tf
@@ -1,0 +1,9 @@
+
+provider "elasticstack" {
+  elasticsearch {}
+  kibana {}
+}
+
+data "elasticstack_kibana_security_role" "example" {
+  name = "sample_role"
+}