+++ examples/resources/elasticstack_kibana_security_role/import.sh
@@ -1,0 +1,1 @@
+terraform import elasticstack_kibana_security_role.example_role <role name>